Updating to latest FlutterFlow output.

+++ lib/pages/meucarrinho/meucarrinho_widget.dart
@@ -555,8 +555,7 @@ class _MeucarrinhoWidgetState extends State<MeucarrinhoWidget>
                                                                     pedidosItem
                                                                         .precoBorda) *
                                                                 (pedidosItem
-                                                                    .quanty
-                                                                    .toDouble());
+                                                                    .quanty);
                                                       });
                                                     },
                                                     child: Icon(
